Init PLT functions separately
The PLT functions were missing since symbol table doesn't link the
undefined symbols to the PLT section.  So the previous logic won't add
them to the function list.

So run objdump for .plt and .plt.got section separately and add the
detected functions.  Note that the first entry in the PLT doesn't have a
symbol so just use the section name.