Add 's' and 'S' keys to save screen shots
The 's' key will create a file named "PREFIX-BINARY-FUNCTION-OFFSET.txt"
where PREFIX can be "das" or "raw" depedning on the mode ('v' key).
The BINARY and FUNCTION would be basenames in case they have file system
separators (e.g. "/").  This file contains exactly the same contents to
the current view (except for the border lines).

The 'S' key will create the same for the whole function.  The OFFSET
part is converted to "full" instead.

Note this only works on the instructions views for now.